xen_version_extra.xsplice: Tool to generate xSplice ELF payload
This change demonstrates how to generate an xSplice ELF payload.

The idea here is that we want to patch in the hypervisor
the 'xen_version_extra' function. To make this even simpler we
would be patching said function with the same function
(TODO: Introduce a new function that returns 'Hello World'.)

To generate this ELF payload file we need:
 - C code (xen_version_extra.c) which has the struct xsplice_*
   implemented. See the xsplice.h header file for the syntax.

 - Hypervisor build-id. We extract that from the xen-syms
   using objcopy and make it part of the ELF payload file.

 - The 'xen_version_extra' code along with its .rodata.

N.B. The xen_version_extra calls the various structures
with a prefix of _NAME. This should be rather a dynamic name
as those names are symbols which will be loaded by the
hypervisor and we do not want collisions.

The linker script file can:
 - Provide a symbol name for the start and end of the
   hypervisor build id. Note that we use -Wl,-build-id=none - this
   is to inhibit the linker from generating an build id for
   the ELF payload.
   TODO: Actally extract the build-id instead of the ELF note!

 - Put the various 'struct xsplice_[sections|patches|code|...]'
   etc in their own sections. The symbolic names of the sections
   (for example xsplice_patches_NAME) are not renamed.
 - Discard .debug* and .comments* sections.
 - Figures out the size of the various .xsplice sections.
 - Aligns the .text, .bss, and .rodata to PAGE_SIZE. This
   should aid us to have PAGE_SIZE pages for each of those
   sections such that we can easily set the right type
   for them (R for .data and .rodata, etc).

It has also has an xen_version_extra specific part - which
is to compute the size of the function it subtracts the
'xen_banner' from 'xen_version_extra' to get the size.
This is a bit of hack as we should figure this out dynamically
(hand waving how). We also end up ingesting the full contents
of the version.o file (including the other functions) - that
should be discarded.

Also if you are curious on the input/output sections
magic the linker does, add these to the GCC line:
  -Wl,-M  -Wl,-t -Wl,-verbose
which are: print linking map, provide trace and be verbose.

*TODO*:
 - Remove the _pad HACK
 - Remove the version_stub.S HACK
 - ELF parser to read xen-syms and have the tool jam it in
   into this file. Actually this file should be generated
   automatically.

Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>