Implement support for applying patches
This is prototype code.

Much of this is implementing a basic version of the Linux kernel module
loader. So the code does:
* Loading of xsplice modules.
* Copying allocated sections into a new executable region of memory.
* Resolving symbols.
* Applying relocations.
* Patching of altinstructions.
* Special handling of bug frames and exception tables.
* Unloading of xsplice modules.

The other main bit is applying and reverting the patches safely. As
implemented, the code is patched with each CPU waiting in the
return-to-guest path (i.e. with no stack) which appears to be the safest
way of patching. It should mean that stack checking is not required.

All of the following should work:
* Applying patches safely.
* Reverting patches safely.
* Replacing patches safely (e.g. reverting any applied patches and
  applying a new patch).
* Bug frames as part of modules. This means adding or changing WARN,
  ASSERT, BUG, and run_in_exception_handler works correctly. Line number
  only changes are ignored.
* Exception tables as part of modules. E.g. wrmsr_safe and copy_to_user
  work correctly in a patch module.
* Hook load and unload functions which run at patch apply and revert
  time respectively.
* Shadow variables. A minimal bit of infrastructure to attach a new
  variable to an existing data structure.

There is no checking that a patch which is loaded is built for the correct hypervisor.